% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaccard_Scoring.R
\name{calculateJaccard}
\alias{calculateJaccard}
\title{Calculate the Jaccard distance}
\usage{
calculateJaccard(a, b)
}
\arguments{
\item{a, b}{character vector, set of gene identifiers.}
}
\value{
The Jaccard distance of the sets.
}
\description{
Calculate the Jaccard distance between two genesets.
}
\examples{
## Mock example showing how the data should look like
a <- c("PDHB", "VARS2")
b <- c("IARS2", "PDHA1")
c <- calculateJaccard(a, b)

## Example using the data available in the package
data(macrophage_topGO_example_small,
     package = "GeDi",
     envir = environment())
genes <- GeDi::prepareGenesetData(macrophage_topGO_example_small)
jaccard <- calculateJaccard(genes[1], genes[2])
}
