% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kappa_Scoring.R
\name{calculateKappa}
\alias{calculateKappa}
\title{Calculate the Kappa distance for two genesets}
\usage{
calculateKappa(a, b, l)
}
\arguments{
\item{a, b}{individual genesets}

\item{l}{the number of unique genes in all input genesets}
}
\value{
the Kappa distance of geneset a and b
}
\description{
Calculate the Kappa distance for two genesets
}
\examples{
a <- c("LARS", "LARS2", "BCAT1")
b <- c("BCAT1", "IARS")
l <- 4
c <- calculateKappa(a, b, l)
}
