% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeDi-data.R
\docType{data}
\name{cluster_macrophage_topGO_example_small}
\alias{cluster_macrophage_topGO_example_small}
\title{Cluster}
\description{
A file containing clusters for the \code{macrophage_topGO_example_small}.
}
\details{
This sample input contains clusters for the
\code{macrophage_topGO_example_small}. Clusters have been calculated on
\code{scores_macrophage_topGO_example_small} using a similarity
threshold of 0.3, a membership threshold of 0.5 and a clustering
threshold of 0.5.
}
\references{
Alasoo, K., Rodrigues, J., Mukhopadhyay, S. et al. Shared
genetic effects on chromatin and gene expression indicate a role
for enhancer priming in immune response. Nat Genet 50, 424–431
(2018). https://doi.org/10.1038/s41588-018-0046-7
}
