% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scIntegration.v113.allow_fixed_B.R
\name{colData.gedi}
\alias{colData.gedi}
\title{Retrieve stored colData}
\usage{
colData.gedi(object)
}
\arguments{
\item{object}{GEDI object}
}
\value{
colData
}
\description{
Retrieve stored colData from GEDI object.
}
