% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\title{Deprecated functions in GeDi}
\arguments{
\item{...}{Ignored arguments.}
}
\value{
All functions throw a warning, with a deprecation message pointing
towards its descendent (if available).
}
\description{
Functions that are on their way to the function afterlife.
Their successors are also listed.
}
\details{
The successors of these functions are likely coming from a renaming of the
functions to more intuitive function names
}
\section{Renaming function with more intuitive names}{

\itemize{
\item \code{\link[=getGenes]{getGenes()}}, now replaced by the more intuitive name
\code{\link[=prepareGenesetData]{prepareGenesetData()}}. The only change in its functionality concerns the
function name.
}
}

\examples{
# try(getGenes())

}
\author{
Annekathrin Nedwed
}
