% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scIntegration.v113.allow_fixed_B.R
\name{dispersion.gedi}
\alias{dispersion.gedi}
\title{Calculate dispersion}
\usage{
dispersion.gedi(object, subsample = 1e+06)
}
\arguments{
\item{object}{GEDI object}

\item{subsample}{subsample}
}
\value{
res dataframe
}
\description{
Calculate the dispersion (variance of observation given the expected value from the model)
}
