% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.checkGenesets}
\alias{.checkGenesets}
\title{Check genesets format}
\usage{
.checkGenesets(
  genesets,
  alt_name_genesets = "Genesets",
  alt_name_genes = "Genes"
)
}
\arguments{
\item{genesets}{a \code{list}, A \code{list} of genesets where each genesets is represented
by \code{list} of genes.}

\item{alt_name_genesets}{character, the name of the column in which the geneset
ids are listed. Defaults to "Genesets".}

\item{alt_name_genes}{character, the name of the column in which the genes
are listed. Defaults to "Genes".}
}
\value{
A validated and formatted genesets data frame.
}
\description{
Check if the input genesets have the expected format for this app
}
