% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.checkPPI}
\alias{.checkPPI}
\title{Check if PPI has right format}
\usage{
.checkPPI(ppi)
}
\arguments{
\item{ppi}{A PPI object}
}
\value{
A \code{data.frame} of the input \code{ppi} with the expected column names
}
\description{
Check if the Protein-Protein-interaction (PPI) has the expected format for
this app
}
