% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.checkScores}
\alias{.checkScores}
\title{Check distance scores format}
\usage{
.checkScores(genesets, distance_scores)
}
\arguments{
\item{genesets}{a \code{list}, A \code{list} of genesets where each genesets is
represented by \code{list} of genes.}

\item{distance_scores}{A \code{\link[Matrix:Matrix]{Matrix::Matrix()}} or  object,
A matrix with numerical (distance) scores.}
}
\value{
A validated and formatted distance_scores \code{\link[Matrix:Matrix]{Matrix::Matrix()}}.
}
\description{
Check if the provided distance scores have the expected format for this app
}
