% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.findSeparator}
\alias{.findSeparator}
\title{Make an educated guess on the separator character}
\usage{
.findSeparator(stringList, sepList = c(",", "\\t", ";", " ", "/"))
}
\arguments{
\item{stringList}{A list of strings}

\item{sepList}{A vector containing the candidates for being identified as
separators. Defaults to \code{c(",", "\t", ";"," ", "/")}.}
}
\value{
A character value, corresponding to the guessed separator. One of ","
(comma), "\\t" (tab), ";" (semicolon)," " (whitespace) or "/"
(backslash).
}
\description{
This function tries to guess which separator was used in a list of delimited
strings.
}
