% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{.getClusterDatatable}
\alias{.getClusterDatatable}
\title{Map each geneset to the cluster it belongs}
\usage{
.getClusterDatatable(cluster, gs_names)
}
\arguments{
\item{cluster}{A \code{list} of clusters}

\item{gs_names}{A vector of geneset names}
}
\value{
A \code{data.frame} mapping each geneset to the cluster(s) it belongs to
}
\description{
Map each geneset to the cluster it belongs and return the information as
a \code{data.frame}
}
\examples{
cluster <- list(c(1:5), c(4:8))
gs_names <- c("a", "b", "c", "d", "e", "f", "g", "h", "i")
df <- getClusterDatatable(cluster, gs_names)
}
