% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.getGenesetDescriptions}
\alias{.getGenesetDescriptions}
\title{Title}
\usage{
.getGenesetDescriptions(genesets)
}
\arguments{
\item{genesets}{a \code{data.frame}, A \code{data.frame} with at least two columns.
One should be called \code{Geneset}, containing the
names/identifiers of the genesets in the data. The second
column should be called \code{Genes} and contains one string of
the genes contained in each geneset.}
}
\value{
a \code{list} of geneset descriptions
}
\description{
Title
}
