% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.getNumberCores}
\alias{.getNumberCores}
\title{Determine the number of cores to use for a function}
\usage{
.getNumberCores(n_cores = NULL)
}
\arguments{
\item{n_cores}{numeric, number of cores to use for the function.
Defaults to \code{Null} in which case the function takes half of
the available cores.}
}
\value{
Number of CPU cores to be used.
}
\description{
Determine the number of CPU cores the scoring functions should use when
computing the distance scores.
}
