% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.sepguesser}
\alias{.sepguesser}
\title{Make an educated guess on the separator character}
\usage{
.sepguesser(file, sep_list = c(",", "\\t", ";", " ", "/"))
}
\arguments{
\item{file}{a character, location of a file to read data from.}

\item{sep_list}{a \code{list}, containing the candidates for being identified as
separators. Defaults to \code{c(",", "\t", ";"," ", "/")}.}
}
\value{
A character, corresponding to the guessed separator. One of ","
(comma), "\\t" (tab), ";" (semicolon)," " (whitespace) or "/"
(backslash).
}
\description{
This function tries to guess which separator was used in a text delimited
file.
}
\references{
See https://github.com/federicomarini/ideal for details on the original
implementation.
}
