% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadPPI.R
\name{downloadPPI}
\alias{downloadPPI}
\title{Title}
\usage{
downloadPPI(species, cachepath = "~/GSD", version = "11.5")
}
\arguments{
\item{species}{the species of the data}

\item{cachepath}{the path to save the cache}

\item{version}{the version of the StringDB database to download from}
}
\value{
the normalized PPI matrix
}
\description{
Title
}
\examples{
\dontrun{}
}
