% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{fuzzy_clustering}
\alias{fuzzy_clustering}
\title{Find cluster from initial seeds}
\usage{
fuzzy_clustering(seeds, threshold)
}
\arguments{
\item{seeds}{A \code{list} of seeds, e.g. determined by seedFinding}

\item{threshold}{numerical, threshold for merging seeds}
}
\value{
A \code{list} of clusters
}
\description{
Find clusters from initial seeds according to (#TODO: David publication linken)
}
\examples{
seeds <- list(c(1:5), c(6:10))
cluster <- fuzzy_clustering(seeds, 0.5)
}
