% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scIntegration.v113.allow_fixed_B.R
\name{getA.gedi}
\alias{getA.gedi}
\title{Return GEDI A}
\usage{
getA.gedi(object)
}
\arguments{
\item{object}{GEDI object}
}
\value{
A matrix
}
\description{
Return the GEDI A projection
}
