% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scIntegration.v113.allow_fixed_B.R
\name{getActivityGradients.gedi}
\alias{getActivityGradients.gedi}
\title{Get Activity gradients}
\usage{
getActivityGradients.gedi(object)
}
\arguments{
\item{object}{GEDI object}
}
\value{
gradient
}
\description{
Return the gradient of all pathways
}
