% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scIntegration.v113.allow_fixed_B.R
\name{getADB.gedi}
\alias{getADB.gedi}
\title{Return GEDI ADB}
\usage{
getADB.gedi(object)
}
\arguments{
\item{object}{GEDI object}
}
\value{
AxDxB matrix
}
\description{
Return the GEDI ADB projection
}
