% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoringGraph.R
\name{getAdjacencyMatrix}
\alias{getAdjacencyMatrix}
\title{Title}
\usage{
getAdjacencyMatrix(distanceMatrix, cutOff)
}
\arguments{
\item{distanceMatrix}{a Distance Matrix containing for each pair the distance}

\item{cutOff}{a cutOff value classifying a threshold for similarity}
}
\description{
Title
}
