% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_PPI.R
\name{getAnnotation}
\alias{getAnnotation}
\title{Get the annotation of a \code{STRINGdb} object}
\usage{
getAnnotation(stringdb)
}
\arguments{
\item{stringdb}{the \code{STRINGdb} object}
}
\value{
A \code{data.frame} mapping \code{STRINGdb} ids to gene names
}
\description{
Get the annotation of a \code{STRINGdb} object, i.e. the aliases of the protein
information
}
\examples{
stringdb <- getStringDB(9606)
stringdb
anno_df <- getAnnotation(stringdb)
}
