% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scIntegration.v113.allow_fixed_B.R
\name{getDB.gedi}
\alias{getDB.gedi}
\title{Return GEDI DB}
\usage{
getDB.gedi(object)
}
\arguments{
\item{object}{GEDI object}
}
\value{
DxB matrix
}
\description{
Return the GEDI DB projection
}
