% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scIntegration.v113.allow_fixed_B.R
\name{getDiffExp.gedi}
\alias{getDiffExp.gedi}
\title{Differential expression}
\usage{
getDiffExp.gedi(object, contrast, include_O = F)
}
\arguments{
\item{object}{GEDI object}

\item{contrast}{a vector representing the sample-variable contrast to use for extracting effect on Qi; must be the same size as aux$L}

\item{include_O}{Logical, indicating whether the effect of contrast on the global offset O}
}
\value{
RDB matrix
}
\description{
Differential expression
}
