% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getGenesIndexed}
\alias{getGenesIndexed}
\title{Assign each gene in the list to its index in the PPI}
\usage{
getGenesIndexed(genes, ppi)
}
\arguments{
\item{genes}{a list of vectors of genes}

\item{ppi}{a Protein-Protein interaction matrix}
}
\value{
a list of vectors of indices
}
\description{
Assign each gene in the list to its index in the PPI
}
