% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_Similarity.R
\name{getGODistanceMatrix}
\alias{getGODistanceMatrix}
\title{Wrapper function}
\usage{
getGODistanceMatrix(
  geneset_ids,
  method = "Wang",
  ontology = "BP",
  species = "org.Hs.eg.db",
  progress = NULL,
  BPPARAM = BiocParallel::SerialParam()
)
}
\arguments{
\item{geneset_ids}{\code{list}, a \code{list} of GO identifiers to score}

\item{method}{character, the method to calculate the GO distance.
See \link[GOSemSim:goSim]{GOSemSim::goSim} measure parameter for possibilities.}

\item{ontology}{character, the ontology to use. See \link[GOSemSim:goSim]{GOSemSim::goSim}
\code{ont} parameter for possibilities.}

\item{species}{character, the species of your data. Indicated as
org.XX.eg.db package from Bioconductor.}

\item{progress}{\code{\link[shiny:Progress]{shiny::Progress()}} object, optional. To track the progress
of the function (e.g. in a Shiny app)}

\item{BPPARAM}{A BiocParallel \code{bpparam} object specifying how parallelization
should be handled. Defaults to \code{\link[BiocParallel:SerialParam-class]{BiocParallel::SerialParam()}}}
}
\value{
A \code{\link[Matrix:Matrix]{Matrix::Matrix()}} with the pairwise GO distance of each
geneset pair.
}
\description{
This is a wrapper function around the goDistance function to match the
names of all the other distance scoring functions.
}
\examples{


## Mock example showing how the data should look like
go_ids <- c("GO:0002503", "GO:0045087", "GO:0019886",
            "GO:0002250", "GO:0001916", "GO:0019885")

similarity <- goDistance(go_ids)

## Example using the data available in the package
data(macrophage_topGO_example_small, package = "GeDi")
go_ids <- macrophage_topGO_example_small$Genesets
\dontrun{
similarity <- goDistance(go_ids)
}
}
