% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pMM_Scoring.R
\name{getInteractionScore}
\alias{getInteractionScore}
\title{Calculate interaction score for two genesets}
\usage{
getInteractionScore(a, b, ppi, maxInteract)
}
\arguments{
\item{a, b}{character vector, set of gene identifiers.}

\item{ppi}{\code{data.frame}, a \code{data.frame} object which contains the
Protein-Protein interactions information. The object is expected
to have three columns, \code{Gene1} and \code{Gene2} which specify the gene
names of the interacting proteins in no particular order
(symmetric interaction) and a column \code{combined_score} which is a
numerical value of the strength of the interaction.}

\item{maxInteract}{numeric, largest value of \code{ppi}.}
}
\value{
The interaction score of the two given genesets.
}
\description{
The function calculates the interaction score for two given genesets.
}
\examples{
a <- c("PDHB", "VARS2", "IARS2")
b <- c("IARS2", "PDHA1")

ppi <- data.frame(
  Gene1 = c("PDHB", "VARS2", "IARS2"),
  Gene2 = c("IARS2", "PDHA1", "CD3"),
  combined_score = c(0.5, 0.2, 0.1)
)
maxInteract <- max(ppi$combined_score)

interaction <- getInteractionScore(a, b, ppi, maxInteract)
}
