% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scIntegration.v113.allow_fixed_B.R
\name{getY.gedi}
\alias{getY.gedi}
\title{Impute Y}
\usage{
getY.gedi(object, logScale = T, rowCentre = T)
}
\arguments{
\item{object}{GEDI object}

\item{logScale}{Whether the imputed values should be on the logarithmic scale}

\item{rowCentre}{Whether the global gene-specific offsets should be removed from imputed values}
}
\value{
Y imputed gene matrix
}
\description{
Impute Y given raw counts, after model fitting is complete
The effect of sample-specific distortions and cell-specific library sizes
will be removed before returning the imputed values.
}
