% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scIntegration.v113.allow_fixed_B.R
\name{getY.var.gedi}
\alias{getY.var.gedi}
\title{Variance of imputed Y}
\usage{
getY.var.gedi(object)
}
\arguments{
\item{object}{GEDI object}
}
\value{
Matrix with variance of imputed Y
}
\description{
Return the variance of the imputed Y (log-scale)
}
