% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scIntegration.v113.allow_fixed_B.R
\name{getZDB.gedi}
\alias{getZDB.gedi}
\title{Return GEDI ZDB}
\usage{
getZDB.gedi(object)
}
\arguments{
\item{object}{GEDI object}
}
\value{
ZxDxB matrix
}
\description{
Return the GEDI ZDB projection
}
