% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goSimilariy.R
\name{goSimilarity}
\alias{goSimilarity}
\title{Calculate similarity of GO terms}
\usage{
goSimilarity(
  geneset_ids,
  method = "Wang",
  ontology = "BP",
  species = "org.Hs.eg.db"
)
}
\arguments{
\item{geneset_ids}{\code{list}, a \code{list} of GO identifiers to score}

\item{method}{character, the method to calculate the GO similarity.
See \link[GOSemSim:goSim]{GOSemSim::goSim} measure parameter for possibilities.}

\item{ontology}{character, the ontology to use. See \link[GOSemSim:goSim]{GOSemSim::goSim}
ont parameter for possibilities.}

\item{species}{character, the species of your data. Indicated as
org.XX.eg.db package from Bioconductor.}
}
\value{
A \code{\link[Matrix:Matrix]{Matrix::Matrix()}} with the pairwise GO similarity of each
geneset pair.
}
\description{
Calculate the pairwise similarity of GO terms
}
\examples{
genesets <- list()
gS <- goSimilarity(genesets)
}
