% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_histogram.R
\name{gs_histogram}
\alias{gs_histogram}
\title{Generate a histogram of geneset size}
\usage{
gs_histogram(
  genes,
  gs_names,
  start = NULL,
  end = NULL,
  binwidth = 5,
  color = "#0092AC"
)
}
\arguments{
\item{genes}{\code{list}, a \code{list} of \code{list} of genes which belong to the genesets
in gs_names}

\item{gs_names}{character vector, names/identifiers of genesets}

\item{start}{numeric, at which number to start the x-axis. Defaults to NULL
meaning the x-axis starts at 0.}

\item{end}{numeric, at which number to end the x-axis. Defaults to NULL,
meaning the size of the largest geneset is the end.}

\item{binwidth}{numeric, size of the individual bins. Defaults to 5.}

\item{color}{character, color to use for the bars of the histogram.
Defaults to #0092AC.}
}
\value{
A \code{ggplot2} histogram with the distribution of the size of the
genesets, i.e. number of genes in the genesets.
}
\description{
Generate a histogram on the size of genesets (i.e. the number of genes
associated with the geneset)
}
\examples{
gs_names <- c("a", "b", "c", "d", "e", "f", "g", "h", "i")
genes <- list(
  c("PDHB", "VARS2"), c("IARS2", "PDHA1"),
  c("AAAS", "ABCE1"), c("ABI1", "AAR2"), c("AATF", "AMFR"),
  c("BMS1", "DAP3"), c("AURKAIP1", "CHCHD1"), c("IARS2"),
  c("AHI1", "ALMS1")
)

p <- gs_histogram(genes, gs_names)
}
