% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadPPI.R
\name{listPPI}
\alias{listPPI}
\title{List all PPIs saved in a Cache}
\usage{
listPPI(cachepath)
}
\arguments{
\item{cachepath}{the path of the cache}
}
\value{
a list of PPIs
}
\description{
List all PPIs saved in a Cache
}
\examples{
\dontrun{}
}
