% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioMart_input_help.R
\name{listSpecies}
\alias{listSpecies}
\title{List valid species for the biomaRt package}
\usage{
listSpecies()
}
\value{
A table with names and descriptions of all the valid species
}
\description{
\code{listSpecies} allows the user to overview all available species for
BioMart Ensembl using the \code{\link[biomaRt:listDatasets]{biomaRt::listDatasets}} function from
the \code{biomaRt} package \insertCite{biomart1,biomart2}{GEDI}. \code{listSpecies} makes it easy to find the
correct species input for the \code{\link{GEDI}} function.
}
\examples{
## Get an overview of all valid species in BioMart
tab <- listSpecies()

## Check if Mus musculus is coverd by BioMart
"mmusculus" \%in\% listSpecies()$species

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{GEDI}} \code{\link{BM_attributes}} \code{\link[biomaRt:listDatasets]{biomaRt::listDatasets}} \code{\link[biomaRt:listAttributes]{biomaRt::listAttributes}}
}
