% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeDi-data.R
\docType{data}
\name{macrophage_KEGG_example}
\alias{macrophage_KEGG_example}
\title{A sample input RData file}
\format{
A data.frame object
}
\description{
A sample input RData file generated from the macrophage dataset.
}
\details{
This sample input contains data from the macrophage package found on
Bioconductor. The exact steps used to generated this file can be
found in the package vignette. The used database for the enrichment
was the KEGG database.
}
\references{
Alasoo, K., Rodrigues, J., Mukhopadhyay, S. et al. Shared
genetic effects on chromatin and gene expression indicate a role
for enhancer priming in immune response. Nat Genet 50, 424–431
(2018). https://doi.org/10.1038/s41588-018-0046-7
}
