% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeDi-data.R
\docType{data}
\name{macrophage_topGO_example_small}
\alias{macrophage_topGO_example_small}
\title{A small sample input RData file}
\format{
A data.frame object
}
\description{
A small sample input RData file generated from the macrophage dataset.
}
\details{
This sample input contains data from the macrophage package found on
Bioconductor. It is a small version of the
\code{macrophage_topGO_example} and only contains the first 50 rows of
this example. It can be used for fast testing of the application.
}
\references{
Alasoo, K., Rodrigues, J., Mukhopadhyay, S. et al. Shared
genetic effects on chromatin and gene expression indicate a role
for enhancer priming in immune response. Nat Genet 50, 424–431
(2018). https://doi.org/10.1038/s41588-018-0046-7
}
