% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadPPI.R
\name{normalizePPI}
\alias{normalizePPI}
\title{Title}
\usage{
normalizePPI(ppi, info)
}
\arguments{
\item{ppi}{the path to a PPI matrix}

\item{info}{the path to the corresponding info file}
}
\value{
a normalized PPI dataframe
}
\description{
Title
}
\examples{
\dontrun{}
}
