% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scIntegration.v113.allow_fixed_B.R
\name{plot.dispersion}
\alias{plot.dispersion}
\title{Plot dispersion}
\usage{
\method{plot}{dispersion}(dispersion)
}
\arguments{
\item{dispersion}{Dispersion dataframe}
}
\value{
ggplot2 object
}
\description{
Plot dispersion
}
