% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scIntegration.v113.allow_fixed_B.R
\name{plot_dispersion}
\alias{plot_dispersion}
\title{Plot dispersion}
\usage{
plot_dispersion(dispersion)
}
\arguments{
\item{dispersion}{Dispersion dataframe}
}
\value{
ggplot2 object
}
\description{
Plot dispersion
}
