% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scIntegration.v113.allow_fixed_B.R
\name{plot_embedding}
\alias{plot_embedding}
\title{Plot 2D embedding}
\usage{
plot_embedding(embedding_mat, colour, randomize = T)
}
\arguments{
\item{embedding_mat}{Embedding}

\item{colour}{vector of variable to plot}

\item{randomize}{Logical. Whether to randomize data before plotting.}
}
\value{
ggplot2 object
}
\description{
Plot a 2D representation (embedding) of cells
}
