% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pMM_Scoring.R
\name{pMMlocal}
\alias{pMMlocal}
\title{Calculates the pMM distance for two genesets}
\usage{
pMMlocal(a, ai, b, bi, alpha, ppi, maxInteract)
}
\arguments{
\item{a, b}{genesets}

\item{ai, bi}{indexed genesets}

\item{alpha}{scaling factor in (0, 1); indicates how much the Protein-Protein interactions are weighted into the distance}

\item{ppi}{a Protein-Protein interaction matrix}

\item{maxInteract}{the maximum value of the Protein-Protein interaction matrix}
}
\value{
pMM distance for two genesets
}
\description{
Calculates the pMM distance for two genesets
}
\examples{
a <- c("PDHB", "VARS2", "IARS2")
ai <- c(1, 2, 3)
b <- c("IARS2", "PDHA2")
bi <- c(4, 3)
ppi <- Matrix::Matrix(0.5, 4, 4)
maxInteract <- 0.5
s <- pMMlocal(a, ai, b, bi, ppi, maxInteract)
}
