% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeDi-data.R
\docType{data}
\name{ppi_macrophage_topGO_example_small}
\alias{ppi_macrophage_topGO_example_small}
\title{PPI}
\description{
A file containing a Protein-Protein Interaction (PPI) \code{data.frame} for the
\code{macrophage_topGO_example_small}.
}
\details{
This sample input contains a PPI for the
\code{macrophage_topGO_example_small}. The PPI has been downloaded using
the functions to download a PPI matrix. Please check out the
vignette for further information.
}
\references{
Alasoo, K., Rodrigues, J., Mukhopadhyay, S. et al. Shared
genetic effects on chromatin and gene expression indicate a role
for enhancer priming in immune response. Nat Genet 50, 424–431
(2018). https://doi.org/10.1038/s41588-018-0046-7
}
