% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadGE.R
\name{ReadGE}
\alias{ReadGE}
\title{Read multiple gene expression datasets}
\usage{
ReadGE(dataFolders, sources, path = NULL, verbose = TRUE)
}
\arguments{
\item{dataFolders}{Character vector with the names of the subdirectories
containing gene expression datasets.}

\item{sources}{Character vector with the data source for each dataset. Valid
sources are \dQuote{affy} or \dQuote{affymetrix}, \dQuote{agilent} and
\dQuote{RNAseq}.}

\item{path}{(Optional) Logical value. If \code{TRUE}, the function prints
descriptive messages about the progress.}

\item{verbose}{(Optional) Logical value. If \code{TRUE}, the function prints
descriptive messages about the progress.}
}
\value{
A list of all transcriptomic datasets that were successfully read where
\code{names(datasets) = dataFolders}.
}
\description{
\code{ReadGE} allows users to read multiple transcriptomic
datasets and store them in a list of \code{data.frames}. A subdirectory
for each dataset contains the data files. The function can read Affymetrix,
Agilent and RNA sequencing data.
}
\details{
\code{ReadGE} reads datasets with different methods depending on the source.
Sample files for Agilent and Affymetrix data are read in alphabetical order.
Samples are presented in the same order on Gene Expression Omnibus (GEO);
\url{https://www.ncbi.nlm.nih.gov/geo/}.

\dQuote{affy}/\dQuote{affymetrix} corresponds to Affymetrix data. \code{ReadGE}
reads Affymetrix data using the \cr\code{\link[affy:ReadAffy]{affy::ReadAffy}} function and
converts it to an expression data table using the \code{\link[gcrma:gcrma]{gcrma::gcrma}}
function. Affymetrix data files have the .CEL or .CEL.gz file extension.

\dQuote{agilent} corresponds to Agilent data. \code{ReadGE} reads Agilent data
using the \code{\link[limma:read.maimages]{limma::read.maimages}} function. Agilent sample files
have the file extension .txt or .txt.gz. The filenames \dQuote{targets.txt} and
\dQuote{annot.txt} are ignored.

\dQuote{RNAseq} corresponds to RNA sequencing data. \code{ReadGE} reads
RNA-seq data using the \code{DESeq2} package. \dQuote{RawCounts.txt} is a
tab-separated table containing RNA-seq count data. This file must contain raw
counts and not transformed or normalised data. \code{ReadGE} transforms the
raw count data using the \code{\link[DESeq2:varianceStabilizingTransformation]{DESeq2::varianceStabilizingTransformation}}
function so the RNA sequencing data can be integrated with microarray data.
\code{ReadGE} does not transform the counts if all datasets are \dQuote{RNAseq}.
}
\examples{
\dontrun{
### From example in README.md
## Three transcriptomic datasets are read
library(GEDI)
# Names of folders with data
dataFolders <- c("RNAseq_data", "affy_data", "agilent_data")
# Which type of data is each dataset
sources <- c("RNAseq", "affy", "agilent")

# Read the data
datasets <- ReadGE(dataFolders, sources,
                   path = "PATH_TO_DATA_FOLDERS")
#> Reading RNAseq_data...
#> Reading affy_data...
#> Reading agilent_data...
}
}
\seealso{
\code{\link{GEDI}}
\code{\link{BatchCorrection}}
\code{\link{VerifyGEDI}}
}
