% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeDi-data.R
\docType{data}
\name{res_macrophage_IFNg_vs_naive}
\alias{res_macrophage_IFNg_vs_naive}
\title{DESeq results}
\description{
A file containing a \code{DESeqResults} for the
\code{macrophage_topGO_example}.
}
\details{
This sample input contains a \code{DESeqResults} for the
\code{macrophage_topGO_examplel}. The object has been generated using the
\code{DESeq} function.
}
\references{
Alasoo, K., Rodrigues, J., Mukhopadhyay, S. et al. Shared
genetic effects on chromatin and gene expression indicate a role
for enhancer priming in immune response. Nat Genet 50, 424–431
(2018). https://doi.org/10.1038/s41588-018-0046-7
}
