% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeDi-data.R
\docType{data}
\name{sample_geneset_broken}
\alias{sample_geneset_broken}
\title{A broken input text file}
\format{
Text file
}
\description{
A broken input text file to test the application
}
\details{
This sample input text file is broken and used for testing the
application.
}
