% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeDi-data.R
\docType{data}
\name{sample_geneset_empty}
\alias{sample_geneset_empty}
\title{An empty input textfile}
\description{
An empty input textfile to test the application
}
\details{
This sample input textfile is empty and used for testing the
application.
}
