% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeDi-data.R
\docType{data}
\name{sample_geneset_small}
\alias{sample_geneset_small}
\title{A small sample input textfile}
\description{
A sample input textfile taken from the GSCluster package, which is reduced
to a smaller number of entries for faster testing of the application-
}
\details{
This sample input textfile contains data from the GSCluster package.
It was taken from the sample_geneset.txt file found on the Github
page of the package and then reduced to a smaller amount of entries
for faster testing of the applcation.
}
\references{
Yoon, S., Kim, J., Kim, SK. et al. GScluster: network-weighted
gene-set clustering analysis. BMC Genomics 20, 352 (2019).
https://doi.org/10.1186/s12864-019-5738-6
}
