% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_Similarity.R
\name{scaleGO}
\alias{scaleGO}
\title{Scaling (distance) scores}
\usage{
scaleGO(
  scores,
  geneset_ids,
  method = "Wang",
  ontology = "BP",
  species = "org.Hs.eg.db",
  BPPARAM = BiocParallel::SerialParam()
)
}
\arguments{
\item{scores}{a \code{\link[Matrix:Matrix]{Matrix::Matrix()}}, a matrix of (distance) scores for the
identifiers in \code{geneset_ids}.}

\item{geneset_ids}{\code{list}, a \code{list} of GO identifiers to score}

\item{method}{character, the method to calculate the GO similarity.
See \link[GOSemSim:goSim]{GOSemSim::goSim} measure parameter for possibilities.}

\item{ontology}{character, the ontology to use. See \link[GOSemSim:goSim]{GOSemSim::goSim}
ont parameter for possibilities.}

\item{species}{character, the species of your data. Indicated as
org.XX.eg.db package from Bioconductor.}

\item{BPPARAM}{A BiocParallelParam object specifying how parallelization
should be handled}
}
\value{
A \code{\link[Matrix:Matrix]{Matrix::Matrix()}} of scaled values.
}
\description{
A method to scale a matrix of distance scores with the GO term similarity
of the associated genesets.
}
\examples{
## Mock example showing how the data should look like
go_ids <- c("GO:0002503", "GO:0045087", "GO:0019886",
            "GO:0002250", "GO:0001916", "GO:0019885")
set.seed(42)
scores <- Matrix::Matrix(stats::runif(36, min = 0, max = 1), 6, 6)
similarity <- scaleGO(scores,
                      go_ids)

## Example using the data available in the package
data(scores_macrophage_topGO_example_small, package = "GeDi")
data(macrophage_topGO_example_small, package = "GeDi")
go_ids <- macrophage_topGO_example_small$Genesets
\dontrun{
scores_scaled <- scaleGO(scores_macrophage_topGO_example_small,
                         go_ids)
}
}
