% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeDi-data.R
\docType{data}
\name{scores_macrophage_topGO_example_small}
\alias{scores_macrophage_topGO_example_small}
\title{Sample scores}
\format{
A sparse matrix (\code{dgCMatrix})
}
\description{
A file containing sample distance scores for the
\code{macrophage_topGO_example_small}.
}
\details{
This sample input contains scores  for the
\code{macrophage_topGO_example_small}. Distance scores have been
calculated using the \code{\link[=getJaccardMatrix]{getJaccardMatrix()}} method.
}
\references{
Alasoo, K., Rodrigues, J., Mukhopadhyay, S. et al. Shared
genetic effects on chromatin and gene expression indicate a role
for enhancer priming in immune response. Nat Genet 50, 424–431
(2018). https://doi.org/10.1038/s41588-018-0046-7
}
