% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pMM_Scoring.R
\name{sumInteraction}
\alias{sumInteraction}
\title{Sums up the protein-protein interacions between genes in two genesets}
\usage{
sumInteraction(x, y, ppi)
}
\arguments{
\item{x, y}{genesets}

\item{ppi}{a Protein-Protein interaction matrix}
}
\value{
sum of the Protein-Protein interactions in the two genesets
}
\description{
Sums up the protein-protein interacions between genes in two genesets
}
\examples{
x <- c(1, 2, 3)
y <- c(3, 1)
ppi <- Matrix::Matrix(0.5, 3, 3)
s <- sumInteraction(x, y, ppi)
}
