% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scIntegration.v113.allow_fixed_B.R
\name{svd.activityGradient.gedi}
\alias{svd.activityGradient.gedi}
\title{SVD activity Gradient}
\usage{
svd.activityGradient.gedi(object, C_index, scale_gradient = NA)
}
\arguments{
\item{object}{GEDI object}

\item{C_index}{the index of the entry in the C matrix whose gradient should be calculated}

\item{scale_gradient}{a scaling factor that determines the magnitude of the gradient vectors.  If NA, the scaling factor will be automatically calculated so that the length of the gradient vector is 20\% of the mean length of cell embedding vectors in the DB projection.}
}
\value{
list with SVD results ( d, u, v, embedding_indices, vectorField_indices )
}
\description{
Calculate the activity gradient
}
