% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scIntegration.v113.allow_fixed_B.R
\name{svd.gedi}
\alias{svd.gedi}
\title{SVD projection of integrated data}
\usage{
svd.gedi(object)
}
\arguments{
\item{object}{GEDI object}
}
\value{
list with SVD results ( d, u, v)
}
\description{
After decomposing the GEDI ZDB projection into U \%\emph{\% S \%}\% t(V), returns S \%*\% t(V)
}
