% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scIntegration.v113.allow_fixed_B.R
\name{svd.joint_vectorField_gradient.gedi}
\alias{svd.joint_vectorField_gradient.gedi}
\title{SVD joint vector Field gradient}
\usage{
svd.joint_vectorField_gradient.gedi(
  object,
  start.cond,
  end.cond,
  C_index,
  scale_cond_vector = 1,
  scale_gradient = NA
)
}
\arguments{
\item{object}{GEDI object}

\item{start.cond}{contrast vector for start condition}

\item{end.cond}{contrast vector for end condition}

\item{C_index}{the index of the entry in the C matrix whose gradient should be calculated.}

\item{scale_cond_vector}{a scaling factor that will be applied to the length of the vectors in the condition-associated vector field}

\item{scale_gradient}{a scaling factor that determines the magnitude of the gradient vectors. If NA, the scaling factor will be automatically calculated so that the length of the gradient vector is 20\% of the mean length of cell embedding vectors in the DB projection.}
}
\value{
list with SVD results ( d, u, v, embedding_indices, vectorField_indices, gradient_indices )
}
\description{
create a umap embedding for cells and their joint vector fields both for sample condition and for pathway activity gradient
}
