% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VerifyGEDI.R
\name{VerifyGEDI}
\alias{VerifyGEDI}
\title{Verify transcriptomic data integration}
\usage{
VerifyGEDI(X, y, batch, model = "logistic", stop.at.score = 1,
           verbose = TRUE)
}
\arguments{
\item{X}{A \code{data.frame} or \code{matrix} of integrated gene expression
data with batch effect removed. Columns contain samples, and rows are genes.}

\item{y}{A character vector containing a status variable for the samples.
For example, healthy/sick or treated/control. \code{y} does not have to
be binary. \code{y} must satisfy: \code{length(y) = ncol(X)}.}

\item{batch}{Character vector that describes which sample is in what batch.
\code{batch} must satisfy: \code{length(batch)=ncol(Data)}.}

\item{model}{(Optional) A character describing which model \code{VerifyGEDI}
uses to predict the status \code{y} in each sample in every batch.
This function currently supports two models; a logistic regression model
(\dQuote{logistic}) and a decision tree classifier (\dQuote{tree}).
\dQuote{logistic} only works if Y contains two classes.}

\item{stop.at.score}{(Optional) The threshold is a numeric value between 0 and 1.
This value determines at what score to force the FSS algorithm to terminate.
The score is how many per cent of the samples are correctly classified.
The default value is 1, meaning that the FSS algorithm will end when the model
makes 0 errors.}

\item{verbose}{(Optional) Logical value. If \code{TRUE}, the function prints
descriptive messages about the progress.}
}
\value{
A list with the following elements:
\describe{
\item{\code{$Feature.idx}}{Indices of rows in \code{X} used as features in the model.}
\item{\code{$Feature.names}}{Names of rows in \code{X} used as features in the model.
These would be Ensembl gene IDs if \code{X} was created using the \code{\link{GEDI}}
function.}
\item{\code{$TrainScores}}{Numeric vector. How the model scores on the training data for each batch.}
\item{\code{$TestScores}}{Numeric vector. Describes how the model scores on each batch
when training on the remaining batches.}
}
}
\description{
\code{VerifyGEDI} allows users to verify if the \code{\link{GEDI}} function
successfully integrated the transcriptomic datasets. Before using this function,
the \code{\link{BatchCorrection}} function must remove the integrated dataset's
batch effect. The transcriptomic data integration is verified using a supervised
machine learning model to predict the samples' status in one batch based on the
remaining batches' samples. The forward stepwise selection (FSS) algorithm picks
which genes to use as features in the chosen model.
}
\details{
There are more genes than samples in gene expression datasets in almost all
cases. Therefore, the \code{VerifyGEDI} function must select some of the genes
as features for the classifier. \code{VerifyGEDI} uses the Forward stepwise
selection (FSS) algorithm to do this.

In each iteration of the FSS algorithm, the chosen model is cross-validated
across all the batches. The model uses one batch as test data and the
remaining batches as training data. This is done for all the batches. The
algorithm uses the lowest test score across all the batches to evaluate which
gene is the best new feature for the model. The FSS algorithm will continue
to do this and add features to the model until the model no longer improves
or that the test score exceeds the threshold determined by the
\code{stop.at.score} argument.
}
\examples{
\dontrun{
### From README.md example
## Continuing example from BatchCorrection()'s example,
## see help("BatchCorrection")

# The gene expression data integration is verified:
res <- VerifyGEDI(X = cData, y = status, batch = batch, model = "logistic")
#> Number of features:  2
#> ENSBTAG00000003530, ENSBTAG00000000476
#> Worst score across batches:  1

# With an accuracy of 100\%, it can be concluded
# that the data integration was successful
}
}
\seealso{
\code{\link{ReadGE}}
\code{\link{GEDI}}
\code{\link{BatchCorrection}}
}
