% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaccard_Scoring.R
\name{calculateJaccard}
\alias{calculateJaccard}
\title{Calculate the Jaccard distance}
\usage{
calculateJaccard(a, b)
}
\arguments{
\item{a, b}{character vector, set of gene identifiers.}
}
\value{
The Jaccard distance of the two sets.
}
\description{
Calculate the Jaccard distance for two given genesets.
}
\examples{
a <- c("PDHB", "VARS2")
b <- c("IARS2", "PDHA1")
c <- calculateJaccard(a, b)
}
