% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kappa_Scoring.R
\name{calculateKappa}
\alias{calculateKappa}
\title{Calculate the Kappa distance for two genesets}
\usage{
calculateKappa(a, b, all_genes)
}
\arguments{
\item{a, b}{individual genesets as two vectors of genes}

\item{all_genes}{the list of all unique genes in all input genesets}
}
\value{
the Kappa distance of geneset a and b
}
\description{
Calculate the Kappa distance for two genesets
}
\examples{
a <- c("LARS", "BCAT1")
b <- c("BCAT1", "IARS")
all_genes <- c("BCAT1", "IARS",  "LARS2")
c <- calculateKappa(a, b, all_genes)
}
