% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kappa_Scoring.R
\name{calculateKappa}
\alias{calculateKappa}
\title{Calculate the Kappa distance}
\usage{
calculateKappa(a, b, all_genes)
}
\arguments{
\item{a, b}{A vector of gene names whose interactions should
be scored.}

\item{all_genes}{A vector of all unique genes in the data set from which \code{a}
and \code{b} are derived.}
}
\value{
The Kappa distance of the two genesets
}
\description{
Calculate the Kappa distance for two given genesets.
}
\examples{
a <- c("PDHB", "VARS2")
b <- c("IARS2", "PDHA1")
all_genes <- c("PDHB", "VARS2", "IARS2", "PDHA1")
c <- calculateKappa(a, b, all_genes)
}
