% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sorensen_dice_Scoring.R
\name{calculateSorensenDice}
\alias{calculateSorensenDice}
\title{Calculate the Sorensen-Dice distance}
\usage{
calculateSorensenDice(a, b)
}
\arguments{
\item{a, b}{character vector, set of gene identifiers.}
}
\value{
The Sorensen-Dice distance of the sets.
}
\description{
Calculate the Sorensen-Dice distance between two genesets.
}
\examples{
#' ## Mock example showing how the data should look like
a <- c("PDHB", "VARS2")
b <- c("IARS2", "PDHA1")
c <- calculateSorensenDice(a, b)

## Example using the data available in the package
data(macrophage_topGO_example_small,
     package = "GeDi",
     envir = environment())
genes <- GeDi::prepareGenesetData(macrophage_topGO_example_small)
sd <- calculateSorensenDice(genes[1], genes[2])
}
