% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{checkInclusion}
\alias{checkInclusion}
\title{Check for subset inclusion}
\usage{
checkInclusion(seeds)
}
\arguments{
\item{seeds}{A \code{list} of sets}
}
\value{
A \code{list} of unique sets
}
\description{
Remove subsets from a given list of sets, i.e. remove sets which are
completely contained in any other larger set in the list.
}
\examples{
seeds <- list(c(1:5), c(2:5), c(6:10))
s <- checkInclusion(seeds)
}
