% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_histogram.R
\name{.buildHistogramData}
\alias{.buildHistogramData}
\title{Prepare data for \code{gsHistogram()}.}
\usage{
.buildHistogramData(genesets, gs_names, start = 0, end = 0)
}
\arguments{
\item{genesets}{a \code{list}, A \code{list} of genesets where each genesets is represented
by \code{list} of genes.}

\item{gs_names}{character vector, Name / identifier of the genesets in
\code{genesets}}

\item{start}{numeric, Optional, describes the minimum gene set size to
include. Defaults to 0.}

\item{end}{numeric, Optional, describes the maximum gene set size to include.
Defaults to 0.}
}
\value{
A \code{data.frame} mapping geneset names to sizes
}
\description{
Prepare the data for the \code{gsHistogram()} by generating a \code{data.frame}
which maps geneset names / identifiers to the size of their size.
}
