% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.checkGenesets}
\alias{.checkGenesets}
\title{Check genesets format}
\usage{
.checkGenesets(genesets)
}
\arguments{
\item{genesets}{a \code{list}, A \code{list} of genesets where each genesets is represented
by \code{list} of genes.}
}
\value{
A validated and formatted genesets data frame.
}
\description{
Check if the input genesets have the expected format for this app
}
