% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.checkGenesets}
\alias{.checkGenesets}
\title{Check if iput genesets have the right format}
\usage{
.checkGenesets(genesets)
}
\arguments{
\item{genesets}{A object of geneset input data}
}
\description{
Check if the input genesets have the expected format of the app
}
