% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.checkGTL}
\alias{.checkGTL}
\title{Check GeneTonic List format}
\usage{
.checkGTL(gtl)
}
\arguments{
\item{gtl}{A \code{GeneTonicList}object generated with
\code{\link[GeneTonic:GeneTonicList]{GeneTonic::GeneTonic_list()}}, containing the functional enrichment
results.}
}
\value{
A validated and renamed geneset \link{data.frame}.
}
\description{
Check if the provided GeneTonic List object has the expected format for the
app and extract the functional enrichment results
}
