% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.filterGenesets}
\alias{.filterGenesets}
\title{Filter Genesets from the input data}
\usage{
.filterGenesets(remove, df_genesets)
}
\arguments{
\item{remove}{\code{list}, a \code{list} of geneset identifiers to be removed from
\code{df_genesets}}

\item{df_genesets}{\code{data.frame}, a \code{data.frame} of the genesets.}
}
\value{
A \code{data.frame} of the input data without the genesets listed in
\code{remove}.
}
\description{
Filter a preselected list of genesets from a \code{data.frame} of genesets
}
