% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.filterGenesets}
\alias{.filterGenesets}
\title{Filter Genesets from the input data}
\usage{
.filterGenesets(remove, df_genesets)
}
\arguments{
\item{remove}{A \code{list} of Geneset identifiers to be removed from the data}

\item{df_genesets}{A \code{data.frame} of the input data}
}
\value{
A \code{data.frame}of the input data without the genesets listed in
\code{remove}
}
\description{
Filter a preselected list of genesets from a \code{data.frame} of genesets
}
