% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.getGenesetDescriptions}
\alias{.getGenesetDescriptions}
\title{Get gene set descriptions}
\usage{
.getGenesetDescriptions(genesets)
}
\arguments{
\item{genesets}{a \code{data.frame}, A \code{data.frame} with at least two columns.
One should be called \code{Geneset}, containing the
names/identifiers of the genesets in the data. The second
column should be called \code{Genes} and contains one string of
the genes contained in each geneset.}
}
\value{
a \code{list} of geneset descriptions
}
\description{
Extracts gene set descriptions from a provided gene set object.
The function prioritizes columns "Term", "Description", or "Genesets" to
find the appropriate descriptions. If any descriptions are duplicated,
the function appends a suffix to make them unique.
}
