% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{fuzzy_clustering}
\alias{fuzzy_clustering}
\title{Find cluster from initial seeds}
\usage{
fuzzy_clustering(seeds, threshold)
}
\arguments{
\item{seeds}{A \code{list} of seeds, e.g. determined by \code{GeDi::seedFinding()}
function}

\item{threshold}{numerical, A threshold for merging seeds}
}
\value{
A \code{list} of clusters
}
\description{
Merge the initially determined seeds to clusters.
}
\examples{
seeds <- list(c(1:5), c(6:10))
cluster <- fuzzy_clustering(seeds, 0.5)
}
\references{
See https://david.ncifcrf.gov/helps/functional_classification.html#clustering
for details on the original implementation
}
