% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEDI.R
\name{GEDI}
\alias{GEDI}
\title{Re-annotate and integrate gene expression datasets}
\usage{
GEDI(datasets, attributes, species = "hsapiens", BioMart = TRUE,
     collapseMethod = "MaxMean", predict.missing = FALSE,
     predict.threshold = 0.8, path=NULL)
}
\arguments{
\item{datasets}{A list of transcriptomic datasets. The \code{\link{ReadGE}}
function can create this list. The \code{GEDI} function integrates all datasets
in the list.}

\item{attributes}{Character vector of BioMart attributes that correspond to
each datasets reporter IDs. The \code{\link{BM_attributes}} function can find
all available BioMart attributes for a given \code{species}.}

\item{species}{(Optional) Character that refers to a species covered by BioMart
Ensembl. The \code{species} must be written with the first letter of the genus
name followed by the specific name, e.g. \emph{Homo sapiens} is \dQuote{hsapiens},
and \emph{Mus musculus} is \dQuote{mmusculus}. The \code{\link{listSpecies}} function
can find all valid species inputs.}

\item{BioMart}{(Optional) Logical value. When \code{TRUE} (default), the
reporter IDs are mapped to Ensembl gene IDs using the \code{biomaRt} package.
If \code{FALSE}, the \code{GEDI} function maps reporter IDs to Ensembl IDs
using a tab-separated annotation table named \dQuote{annot.txt} for each dataset.
The reporter IDs must be in the first column of the table. The second column holds
Ensembl gene IDs. The second columns name must correspond to the BioMart attribute,
e.g. \dQuote{ensembl_gene_id}.}

\item{collapseMethod}{(Optional) Character that determines which method the
\code{\link[WGCNA:collapseRows]{WGCNA::collapseRows}} function uses to collapse two rows
mapped to the same Ensembl gene ID. Valid options are: \code{MaxMean} (default),
\code{MinMean}, \code{absMaxMean}, \code{absMinMean}, \code{Average},
\code{maxRowVariance} and \code{ME}. See \code{\link[WGCNA:collapseRows]{WGCNA::collapseRows}}
documentation for more info.}

\item{predict.missing}{(Optional) Logical value. When \code{TRUE}, \code{GEDI}
predicts missing expression values for a gene if the percentage of known values
is higher than the \code{predict.threshold}. This does not currently work for
count data.}

\item{predict.threshold}{(Optional) Numeric argument between 0 and 1. This
argument is ignored if \code{predict.missing = FALSE}.}

\item{path}{(Optional) (Optional) Path to the directory where all data folders
are located. The data folders store annotation files if there are any.
The current working directory is the default. The names of the data folders
are \code{names(datasets)}.}
}
\value{
A \code{Data.frame} containing columns from all datasets. All rows are mapped
to Ensembl gene IDs. Rows with missing data have been removed from the output.
}
\description{
\code{GEDI} allows users to integrate transcriptomic datasets
into one data table containing samples from all the datasets. The first step
is to map the reporter IDs to Ensembl gene IDs. The second step is to collapse
rows mapped to the same Ensembl gene ID. The third step is to merge the
datasets. Optionally, \code{GEDI} can fill in missing data using a linear
regression model.
}
\details{
The \code{GEDI} function maps reporter IDs of the datasets to Ensembl gene IDs
using the \code{biomaRt} package \insertCite{biomart1,biomart2}{GEDI}.
However, if the BioMart attribute for a dataset is invalid or missing, an
annotation table called \dQuote{annot.txt} is needed;
see the \code{BioMart} argument. If the second column does not contain Ensembl
gene IDs, then it must contain another type of reporter ID with a known BioMart
attribute.
The annotation file must be in the same subdirectory as the corresponding
dataset. The subdirectory's name must correspond to the dataset's name; see
\code{names(datasets)}. The annotation file is also used if the user chooses
not to use BioMart.

Multiple reporter IDs can map to the same Ensembl gene ID. Therefore, these
reporter IDs' expression values must be collapsed using the
\code{\link[WGCNA:collapseRows]{WGCNA::collapseRows}} function. The method used to collapse
the rows can be specified with the \code{collapseMethod} argument. After this,
the datasets are merged.

If \code{predict.missing = TRUE}, then the missing expression values are
predicted for all Ensembl gene IDs where the percentage of known expression
values exceeds the \code{predict.threshold} argument. \code{GEDI} uses a
linear regression model to predict the missing expression values. The linear
regression model uses only one feature; the gene most correlated with the gene
with missing values. Only one feature is chosen to get a quick estimate.
However, it can still be time-consuming if many rows have missing values.
}
\examples{
\dontrun{
### From example in README.md
## Continuing example from ReadGE()'s example, see help("ReadGE")
# Biomart attributes defined for the three datasets
# The species is Bos taurus
attr <- c("ensembl_gene_id", affy_bovine, NA)

# There are no attribute for the third dataset
# An annotation file (annot.txt) are used
# It is located in the datafolder for the dataset

dat <- GEDI(datasets, attributes = attr, BioMart = TRUE,
           species = "btaurus", path = PATH_TO_DATA_FOLDERS)
           #> Connecting to BioMart...
           #> Invalid BioMart attribute, NA in agilent_data
           #> Reading annotation file ('annot.txt') for agilent_data
dim(dat)
#> [1] 8328   20

# The integrated dataset has 20 samples with expression values
# for 8328 genes
}
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{ReadGE}}
\code{\link{BatchCorrection}}
\code{\link{VerifyGEDI}}
}
