% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadPPI.R
\name{getAnnotation}
\alias{getAnnotation}
\title{Get the annotation of a \code{STRINGdb} object}
\usage{
getAnnotation(stringdb)
}
\arguments{
\item{stringdb}{the \code{STRINGdb} object to get the annotation from}
}
\value{
a \code{data.frame} which maps \code{STRING_ids} to \code{alias}, which are gene names
}
\description{
Get the annotation of a \code{STRINGdb} object
}
\examples{
string_db <- getStringDB(9606)
string_db
anno_df <- getAnnotation(string_db)
}
