% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadPPI.R
\name{getAnnotation}
\alias{getAnnotation}
\title{Get the annotation of a \link{STRINGdb} object}
\usage{
getAnnotation(stringdb)
}
\arguments{
\item{stringdb}{the \link{STRINGdb} object}
}
\value{
A \code{data.frame} mapping \link{STRINGdb} ids to gene names
}
\description{
Get the annotation of a \link{STRINGdb} object, i.e. the aliases of the protein
information
}
\examples{
string_db <- getStringDB(9606)
string_db
anno_df <- getAnnotation(string_db)
}
