% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getGenes}
\alias{getGenes}
\title{Split the genesets into a list of vectors of genes, one for each geneset}
\usage{
getGenes(genesets)
}
\arguments{
\item{genesets}{a list of the genesets}
}
\value{
a list of vectors of genes, on for each geneset
}
\description{
Split the genesets into a list of vectors of genes, one for each geneset
}
\examples{
genesets <- list(c("PDHB", "VARS2", "IARS2"), c("IARS2", "PDHA2"))
genes <- getGenes(genesets)
}
