% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadPPI.R
\name{getId}
\alias{getId}
\title{Get NCBI ID of a species}
\usage{
getId(species, version = "11.5")
}
\arguments{
\item{species}{the species you want to get the NCBI ID of}

\item{version}{}
}
\value{
the NCBI ID of your searched species
}
\description{
Get NCBI ID of a species
}
\examples{
id <- getId(species = "Homo Sapiens")
Title

}
