% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadPPI.R
\name{getId}
\alias{getId}
\title{Get ID of a species}
\usage{
getId(species, version = "11.5")
}
\arguments{
\item{species}{character, the species of your input data}

\item{version}{character, the version of STRING you want to use, defaults to
11.5, the current version of STRING}
}
\value{
a character of the NCBI ID of \code{species}
}
\description{
Get the NCBI ID of a species
}
\examples{
species <- "Homo sapiens"
id <- getId(species = species)
}
