% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadPPI.R
\name{getId}
\alias{getId}
\title{Translates the species name into the NCBI ID}
\usage{
getId(species)
}
\arguments{
\item{species}{the species of the data to download}
}
\value{
the NCBI id of the species
}
\description{
Translates the species name into the NCBI ID
}
\examples{
species <- "Homo sapiens"
getId(species)
}
