% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pMM_Scoring.R
\name{getInteractionScore}
\alias{getInteractionScore}
\title{Calculates the interaction score for two genesets}
\usage{
getInteractionScore(a, ai, b, bi, ppi, maxInteract)
}
\arguments{
\item{a, b}{genesets}

\item{ai, bi}{indexed genesets}

\item{ppi}{a Protein-Protein interaction matrix}

\item{maxInteract}{the maximum value in the Protein-Protein interaction matrix}
}
\value{
the interaction score for two genesets
}
\description{
Calculates the interaction score for two genesets
}
\examples{
a <- c("PDHB", "VARS2", "IARS2")
ai <- c(1, 2, 3)
b <- c("IARS2", "PDHA2")
bi <- c(4, 3)
ppi <- Matrix::Matrix(0.5, 4, 4)
maxInteract <- 0.5
s <- getInteractionScore(a, ai, b, bi, ppi, maxInteract)
}
