% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pMM_Scoring.R
\name{getInteractionScore}
\alias{getInteractionScore}
\title{Calculate interaction score for two genesets}
\usage{
getInteractionScore(a, b, ppi, maxInteract)
}
\arguments{
\item{a, b}{A vector of gene names whose interactions should
be used.}

\item{ppi}{A \code{data.frame} object which contains the Protein-Protein
interactions.
The object has three columns, \code{from} and \code{to} which
specify the gene names of the interacting proteins in no
particular order (symmetric interaction) and a column
\code{combined_score} which is a numerical value of the strength of
the interaction.}

\item{maxInteract}{The maximum value of the \code{combined_score} column of
the \code{ppi}.}
}
\value{
The interaction score of the two given genesets.
}
\description{
The function calculates the interaction score for two given genesets.
}
\examples{
a <- c("PDHB", "VARS2")
b <- c("IARS2", "PDHA1")

ppi <- data.frame(from = c("PDHB", "VARS2"),
                  to = c("IARS2", "PDHA1"),
                  combined_score = c(0.5, 0.2))
maxInteract <- max(ppi$combined_score)

interaction <- getInteractionScore(a, b, ppi, maxInteract)
}
