% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaccard_Scoring.R
\name{getJaccardMatrix}
\alias{getJaccardMatrix}
\title{Get Matrix of Jaccard distances}
\usage{
getJaccardMatrix(genesets, progress = NULL)
}
\arguments{
\item{genesets}{A \code{list} of genesets (each geneset is represented by a \code{list}
of the corresponding genes).}

\item{progress}{An optional \code{\link[shiny:Progress]{shiny::Progress()}} object to track the progress
progress of the function in the app.}
}
\value{
A \code{\link[Matrix:Matrix]{Matrix::Matrix()}} with the pairwise Jaccard distance of each
geneset pair.
}
\description{
Calculate the Jaccard distance of all combinations of genesets in a given data
set of genesets.
}
\examples{
genesets <- list(list("PDHB", "VARS2"), list("IARS2", "PDHA1"))
m <- getJaccardMatrix(genesets)
}
