% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kappa_Scoring.R
\name{getKappaMatrix}
\alias{getKappaMatrix}
\title{Get Matrix of Kappa distances}
\usage{
getKappaMatrix(genesets, progress = NULL)
}
\arguments{
\item{genesets}{A \code{list} of genesets (each geneset is represented by a \code{list}
of the corresponding genes).}

\item{progress}{An optional \code{\link[shiny:Progress]{shiny::Progress()}} object to track the progress
progress of the function in the app.}
}
\value{
A \code{\link[Matrix:Matrix]{Matrix::Matrix()}} with the pairwise Kappa distance of each
geneset pair.
}
\description{
Calculate the Kappa distance of all combinations of genesets in a given data
set of genesets.
}
\examples{
genesets <- list(list("PDHB", "VARS2"), list("IARS2", "PDHA1"))
m <- getKappaMatrix(genesets)
}
