% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kappa_Scoring.R
\name{getKappaMatrix}
\alias{getKappaMatrix}
\title{Get Matrix of Kappa distances}
\usage{
getKappaMatrix(genesets, progress = NULL, n_cores = NULL)
}
\arguments{
\item{genesets}{\code{list}, a \code{list} of genesets (each geneset is represented by
a \code{list} of the corresponding genes).}

\item{progress}{\code{\link[shiny:Progress]{shiny::Progress()}} object, optional. To track the progress
of the function (e.g. in a Shiny app)}

\item{n_cores}{numeric, number of cores to use for the function.
Defaults to \code{NULL} in which case the function takes half of
the available cores (see function .detectNumberCores(n_cores)).}
}
\value{
A \code{\link[Matrix:Matrix]{Matrix::Matrix()}} with the pairwise Kappa distance of each
geneset pair. The matrix is symmetrical with values between 0 and 1,
where 0 indicates the smallest distance (identical genesets) and
1 indicates two disjoint sets.
}
\description{
Calculate the Kappa distance of all combinations of genesets in a given data
set of genesets. The Kappa distance is normalized to the (0, 1) interval.
}
\examples{
genesets <- list(list("PDHB", "VARS2"), list("IARS2", "PDHA1"))
m <- getKappaMatrix(genesets, n_cores = 1)
}
