% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meetmin_Scoring.R
\name{getMeetMinMatrix}
\alias{getMeetMinMatrix}
\title{Get Matrix of Meet-Min distances}
\usage{
getMeetMinMatrix(genesets)
}
\arguments{
\item{genesets}{A \code{list} of genesets (each geneset is represented by a \code{list}
of the corresponding genes).}
}
\value{
A \code{\link[Matrix:Matrix]{Matrix::Matrix()}} with the pairwise Meet-Min distance of each
geneset pair.
}
\description{
Calculate the Meet-Min distance of all combinations of genesets in a given
data set of genesets.
}
\examples{
genesets <- list(list("PDHB", "VARS2"), list("IARS2", "PDHA1"))
m <- getMeetMinMatrix(genesets)
}
