% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meetmin_Scoring.R
\name{getMeetMinMatrix}
\alias{getMeetMinMatrix}
\title{Calculate the Meet-Min distance of all geneset combinations}
\usage{
getMeetMinMatrix(genesets, geneset_names)
}
\arguments{
\item{genesets}{a list of the genesets to score}

\item{geneset_names}{a list of the names/ ids of the genesets}
}
\value{
a \code{\link[Matrix:Matrix]{Matrix::Matrix()}} with the pairwise Meet-Min distance of each geneset pair
}
\description{
Calculate the Meet-Min distance of all geneset combinations
}
\examples{
genesets <- list(c("PDHB", "VARS2", "IARS2"), c("IARS2", "PDHA2"))
geneset_names <- list("A", "B")
m <- getMeetMinMatrix(genesets, geneset_names)
}
