% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pMM_Scoring.R
\name{getpMMMatrix}
\alias{getpMMMatrix}
\title{Calculate the pMM distance of all geneset combinations}
\usage{
getpMMMatrix(genes, ppi, geneset_names, alpha = 1)
}
\arguments{
\item{genes}{a list of  vectors of genes (the genesets)}

\item{ppi}{a Protein-Protein interaction matrix}

\item{geneset_names}{a list of names/ids of the genesets}

\item{alpha}{scaling factor in (0, 1); indicates how much the Protein-Protein interactions are weighted into the distance}
}
\value{
a \code{\link[Matrix:Matrix]{Matrix::Matrix()}} with the pairwise pMM distance of each geneset pair
}
\description{
Calculate the pMM distance of all geneset combinations
}
\examples{
genes <- list(c("PDHB", "VARS2", "IARS2"), c("IARS2", "PDHA2"))
geneset_names <- list("A", "B")
ppi <- Matrix::Matrix(0.5, 4, 4)
s <- getpMMMatrix(genes, ppi, geneset_names)
}
