% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pMM_Scoring.R
\name{getpMMMatrix}
\alias{getpMMMatrix}
\title{Calculate the pMM distance of all geneset combinations}
\usage{
getpMMMatrix(genesets, ppi, alpha = 1)
}
\arguments{
\item{genesets}{a list of the genesets to score}

\item{ppi}{a Protein-Protein interaction matrix}

\item{alpha}{scaling factor in (0, 1); indicates how much the Protein-Protein interactions are weighted into the distance}
}
\value{
a \code{\link[Matrix:Matrix]{Matrix::Matrix()}} with the pairwise pMM distance of each geneset pair
}
\description{
Calculate the pMM distance of all geneset combinations
}
\examples{
genesets <- list(c("PDHB", "VARS2", "IARS2"), c("IARS2", "PDHA2"))
ppi <- Matrix::matrix(0.5, 4, 4)
s <- pMM(genesets, ppi)
}
