% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pMM_Scoring.R
\name{getpMMMatrix}
\alias{getpMMMatrix}
\title{Calculate the pMM distance}
\usage{
getpMMMatrix(genes, ppi, alpha = 1, progress = NULL, n_cores = NULL)
}
\arguments{
\item{genes}{\code{list}, a \code{list} of genesets (each geneset is represented by
a \code{list} of the corresponding genes).}

\item{ppi}{\code{data.frame}, a \code{data.frame} object which contains the
Protein-Protein interactions information. The object is expected
to have three columns, \code{Gene1} and \code{Gene2} which specify the gene
names of the interacting proteins in no particular order
(symmetric interaction) and a column \code{combined_score} which is a
numerical value of the strength of the interaction.}

\item{alpha}{numeric, scaling factor in (0, 1) which indicates how much the
Protein-Protein interactions are weighted into the final score.
Defaults to 1.}

\item{progress}{\code{\link[shiny:Progress]{shiny::Progress()}} object, optional. To track the progress
of the function (e.g. in a Shiny app)}

\item{n_cores}{numeric, number of cores to use for the function.
Defaults to \code{NULL} in which case the function takes half of
the available cores (see function .detectNumberCores(n_cores)).}
}
\value{
A \code{\link[Matrix:Matrix]{Matrix::Matrix()}} with the pairwise pMM distance of each
geneset pair.
}
\description{
Calculate the pMM distance of all combinations of genesets in a given data
set of genesets.
}
\examples{
genes <- list(c("PDHB", "VARS2"), c("IARS2", "PDHA1"))

ppi <- data.frame(
  Gene1 = c("PDHB", "VARS2"),
  Gene2 = c("IARS2", "PDHA1"),
  combined_score = c(0.5, 0.2)
)

pMM <- getpMMMatrix(genes, ppi, n_cores = 1)
}
