% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pMM_Scoring.R
\name{getpMMMatrix}
\alias{getpMMMatrix}
\title{Calculate the pMM distance}
\usage{
getpMMMatrix(genesets, ppi, alpha = 1, progress = NULL, n_cores = NULL)
}
\arguments{
\item{genesets}{a \code{list}, A \code{list} of genesets where each genesets is represented
by \code{list} of genes.}

\item{ppi}{a \code{data.frame}, Protein-protein interaction (PPI) network data frame.
The object is expected to have three columns, \code{Gene1} and \code{Gene2}
which specify the gene names of the interacting proteins in no
particular order (symmetric interaction) and a column
\code{combined_score} which is a numerical value of the strength of
the interaction.}

\item{alpha}{numeric, Scaling factor for controlling the influence of the
interaction score. Defaults to 1.}

\item{progress}{a \code{\link[shiny:Progress]{shiny::Progress()}} object, Optional progress bar object
to track the progress of the function (e.g. in a Shiny app).}

\item{n_cores}{numeric, Optional number of CPU cores to use for parallel
processing. Defaults to \code{NULL} in which case the function
takes half of the available cores (see
\code{.detectNumberCores()} function).}
}
\value{
A \code{\link[Matrix:Matrix]{Matrix::Matrix()}} with pMM distance rounded to 2 decimal places.
}
\description{
Calculate the pMM distance of all combinations of genesets in a given data
set of genesets.
}
\examples{
genesets <- list(c("PDHB", "VARS2"), c("IARS2", "PDHA1"))

ppi <- data.frame(
  Gene1 = c("PDHB", "VARS2"),
  Gene2 = c("IARS2", "PDHA1"),
  combined_score = c(0.5, 0.2)
)

pMM <- getpMMMatrix(genesets, ppi, n_cores = 1)
}
\references{
See https://doi.org/10.1186/s12864-019-5738-6 for details on the original
implementation.
}
