% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sorensen_dice_Scoring.R
\name{getSorensenDiceMatrix}
\alias{getSorensenDiceMatrix}
\title{Get Matrix of Sorensen-Dice distances}
\usage{
getSorensenDiceMatrix(genesets, progress = NULL, n_cores = NULL)
}
\arguments{
\item{genesets}{a \code{list}, A \code{list} of genesets where each genesets is represented
by \code{list} of genes.}

\item{progress}{a \code{\link[shiny:Progress]{shiny::Progress()}} object, Optional progress bar object
to track the progress of the function (e.g. in a Shiny app).}

\item{n_cores}{numeric, Optional number of CPU cores to use for parallel
processing. Defaults to \code{NULL} in which case the function
takes half of the available cores (see
\code{.detectNumberCores()} function).}
}
\value{
A \code{\link[Matrix:Matrix]{Matrix::Matrix()}} with Sorensen-Dice distance rounded to 2 decimal
places.
}
\description{
Calculate the Sorensen-Dice distance of all combinations of genesets in a given data
set of genesets.
}
\examples{
## Mock example showing how the data should look like
genesets <- list(list("PDHB", "VARS2"), list("IARS2", "PDHA1"))
m <- getSorensenDiceMatrix(genesets, n_cores = 1)

## Example using the data available in the package
data(macrophage_topGO_example_small,
     package = "GeDi",
     envir = environment())
genes <- GeDi::getGenes(macrophage_topGO_example_small)
sd_matrix <- getSorensenDiceMatrix(genes, n_cores = 1)
}
