% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadPPI.R
\name{getStringDB}
\alias{getStringDB}
\title{Title}
\usage{
getStringDB(
  species,
  version = "11.5",
  score_threshold = 0,
  input_directory = ""
)
}
\arguments{
\item{score_threshold}{}
}
\description{
Title
}
