% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadPPI.R
\name{getStringDB}
\alias{getStringDB}
\title{Get the STRING entry of a species}
\usage{
getStringDB(species, version = "11.5", score_threshold = 0)
}
\arguments{
\item{species}{numeric, the NCBI ID of the species of interest}

\item{version}{character, the STRINGdb version to use, defaults to 11.5
(current version)}

\item{score_threshold}{numeric, a score threshold to cut the retrieved
interactions, defaults to 0 (all interactions)}
}
\value{
a \link{STRINGdb} object \code{species}
}
\description{
Get the respective \link{STRINGdb} object of your species of interest
}
\examples{
species <- getId(species = "Homo sapiens")
string_db <- getStringDB(as.numeric(species))
}
