% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goSimilariy.R
\name{goSimilarity}
\alias{goSimilarity}
\title{Calculate similarity of GO terms}
\usage{
goSimilarity(
  geneset_ids,
  method = "Wang",
  ontology = "BP",
  species = "org.Hs.eg.db",
  progress = NULL,
  n_cores = NULL
)
}
\arguments{
\item{geneset_ids}{\code{list}, a \code{list} of GO identifiers to score}

\item{method}{character, the method to calculate the GO similarity.
See \link[GOSemSim:goSim]{GOSemSim::goSim} measure parameter for possibilities.}

\item{ontology}{character, the ontology to use. See \link[GOSemSim:goSim]{GOSemSim::goSim}
\code{ont} parameter for possibilities.}

\item{species}{character, the species of your data. Indicated as
org.XX.eg.db package from Bioconductor.}

\item{progress}{\code{\link[shiny:Progress]{shiny::Progress()}} object, optional. To track the progress
of the function (e.g. in a Shiny app)}

\item{n_cores}{numeric, number of cores to use for the function.
Defaults to \code{NULL} in which case the function takes half of
the available cores (see function .detectNumberCores(n_cores)).}
}
\value{
A \code{\link[Matrix:Matrix]{Matrix::Matrix()}} with the pairwise GO similarity of each
geneset pair.
}
\description{
Calculate the pairwise similarity of GO terms
}
\examples{
data(macrophage_topGO_example_small, package = "GeDi")
go_ids <- macrophage_topGO_example_small$Genesets
similarity <- goSimilarity(go_ids,
                           n_cores = 1)
}
