% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goSimilariy.R
\name{goSimilarity}
\alias{goSimilarity}
\title{Calculate the GO Similarity of all geneset combinations}
\usage{
goSimilarity(
  genesets,
  method = "Wang",
  ontology = "BP",
  species = "org.Hs.eg.db"
)
}
\arguments{
\item{genesets}{a list of vectors of genes to score, one vector for each geneset}

\item{method}{the method to calculate the GO similarity see \link{GOSemSim} measure parameter}

\item{ontology}{the ontology of GO to use, see \link{GOSemSim} ont parameter}

\item{species}{the species of your data, indicated as org.XX.eg.db package from Bioconductor}
}
\value{
a \code{\link[Matrix:Matrix]{Matrix::Matrix()}} with the pairwise GO similarity of each geneset pair
}
\description{
Calculate the GO Similarity of all geneset combinations
}
\examples{
\dontrun{
genesets <- list(c("PDHB", "VARS2", "IARS2"), c("IARS2", "PDHA2"))
go <- goSimilarity(genesets)
}
}
