% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pMM_Scoring.R
\name{pMMlocal}
\alias{pMMlocal}
\title{Calculate local pMM distance}
\usage{
pMMlocal(a, b, ppi, maxInteract, alpha = 1)
}
\arguments{
\item{a, b}{A vector of gene names whose interactions should
be scored.}

\item{ppi}{A \code{data.frame} object which contains the Protein-Protein
interactions.
The object has three columns, \code{Gene1} and \code{Gene2} which
specify the gene names of the interacting proteins in no
particular order (symmetric interaction) and a column
\code{combined_score} which is a numerical value of the strength of
the interaction.}

\item{maxInteract}{The maximum value of the \code{combined_score} column of
the \code{ppi}.}

\item{alpha}{A scaling factor (between 0 and 1) which indicates how much the
Protein-Protein interactions are weighted into the final score.
Defaults to 1.}
}
\value{
The pMM distance for two genesets.
}
\description{
Calculate the local pMM distance of two genesets.
}
\examples{
a <- c("PDHB", "VARS2")
b <- c("IARS2", "PDHA1")

ppi <- data.frame(
  Gene1 = c("PDHB", "VARS2"),
  Gene2 = c("IARS2", "PDHA1"),
  combined_score = c(0.5, 0.2)
)
maxInteract <- max(ppi$combined_score)

pMM_score <- pMMlocal(a, b, ppi, maxInteract)
}
