% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pMM_Scoring.R
\name{pMMlocal}
\alias{pMMlocal}
\title{Calculate local pMM distance}
\usage{
pMMlocal(a, b, ppi, maxInteract, alpha = 1)
}
\arguments{
\item{a, b}{character vector, set of gene identifiers.}

\item{ppi}{\code{data.frame}, a \code{data.frame} object which contains the
Protein-Protein interactions information. The object is expected
to have three columns, \code{Gene1} and \code{Gene2} which specify the gene
names of the interacting proteins in no particular order
(symmetric interaction) and a column \code{combined_score} which is a
numerical value of the strength of the interaction.}

\item{maxInteract}{numeric, largest value of \code{ppi}.}

\item{alpha}{numeric, scaling factor in (0, 1) which indicates how much the
Protein-Protein interactions are weighted into the final score.
Defaults to 1.}
}
\value{
The pMM distance for two sets.
}
\description{
Calculate the local pMM distance of two genesets.
}
\examples{
a <- c("PDHB", "VARS2")
b <- c("IARS2", "PDHA1")

ppi <- data.frame(
  Gene1 = c("PDHB", "VARS2"),
  Gene2 = c("IARS2", "PDHA1"),
  combined_score = c(0.5, 0.2)
)
maxInteract <- max(ppi$combined_score)

pMM_score <- pMMlocal(a, b, ppi, maxInteract)
}
