% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeDi-data.R
\docType{data}
\name{sample_geneset}
\alias{sample_geneset}
\title{A sample input textfile}
\description{
A sample input textfile taken from the GSCluster package
}
\details{
This sample input textfile contains data from the GSCluster package.
It is identical to the sample_geneset.txt file found on the Github
page of the package.
}
\references{
Yoon, S., Kim, J., Kim, SK. et al. GScluster: network-weighted
gene-set clustering analysis. BMC Genomics 20, 352 (2019).
https://doi.org/10.1186/s12864-019-5738-6
}
