% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeDi-data.R
\docType{data}
\name{sample_geneset_broken}
\alias{sample_geneset_broken}
\title{A broken input textfile}
\description{
A broken input textfile to test the application
}
\details{
This sample input textfile is broken and used for testing the
application.
}
