% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goSimilariy.R
\name{scaleGO}
\alias{scaleGO}
\title{Scaling computed distances with GO similarity scores}
\usage{
scaleGO(
  scores,
  genesets,
  method = "Wang",
  ontology = "BP",
  species = "org.Hs.eg.db"
)
}
\arguments{
\item{scores}{a \code{\link[Matrix:Matrix]{Matrix::Matrix()}} with the pairwise distances of each geneset pair}

\item{genesets}{a list of vectors of genes to score, one vector for each geneset}

\item{method}{the method to calculate the GO similarity see \link{GOSemSim} measure parameter}

\item{ontology}{the ontology of GO to use, see \link{GOSemSim} ont parameter}

\item{species}{the species of your data, indicated as org.XX.eg.db package from Bioconductor}
}
\description{
Scaling computed distances with GO similarity scores
}
