% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pMM_Scoring.R
\name{sumInteraction}
\alias{sumInteraction}
\title{Sum up Protein-Protein interactions between two genesets}
\usage{
sumInteraction(a, b, ppi)
}
\arguments{
\item{a, b}{A vector of gene names whose interactions should
be scored.}

\item{ppi}{A \code{data.frame} object which contains the Protein-Protein
interactions.
The object has three columns, \code{from} and \code{to} which
specify the gene names of the interacting proteins in no
particular order (symmetric interaction) and a column
\code{combined_score} which is a numerical value of the strength of
the interaction.}
}
\value{
The sum of the Protein-Protein interactions.
}
\description{
The function sums up the strength of the Protein-Protein interactions of
all genes in the given genesets.
}
\examples{
a <- c("PDHB", "VARS2")
b <- c("IARS2", "PDHA1")

ppi <- data.frame(from = c("PDHB", "VARS2"),
                  to = c("IARS2", "PDHA1"),
                  combined_score = c(0.5, 0.2))

score <- sumInteraction(a, b, ppi)
}
