% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scIntegration.v113.allow_fixed_B.R
\name{svd.vectorField.gedi}
\alias{svd.vectorField.gedi}
\title{SVD for vector fields}
\usage{
svd.vectorField.gedi(object, start.cond, end.cond)
}
\arguments{
\item{object}{GEDI object}

\item{start.cond}{contrast vector for start condition}

\item{end.cond}{contrast vector for end condition}
}
\value{
list with SVD results ( d, u, v)
}
\description{
Calculate svd for cells and their vector fields.
}
